Saves

+++ Controllers/AdminController.cs
@@ -33,13 +33,13 @@ namespace ASP_SPR311.Controllers
         [HttpPost]
         public JsonResult AddProduct(ProductFormModel formModel)
         {
-            double price;
+            decimal price;
             try { 
-                price = double.Parse(formModel.Price, System.Globalization.CultureInfo.InvariantCulture); 
+                price = decimal.Parse(formModel.Price, System.Globalization.CultureInfo.InvariantCulture); 
             }
             catch
             {
-                price = double.Parse(formModel.Price.Replace(',', '.'), System.Globalization.CultureInfo.InvariantCulture);
+                price = decimal.Parse(formModel.Price.Replace(',', '.'), System.Globalization.CultureInfo.InvariantCulture);
             }
             Product product = new()
             {
